# docs/swagger.json
{"basePath": "/api", "consumes": ["application/json"], "definitions": {"Site": {"properties": {"url": {"description": "URL", "type": "string"}}, "required": ["url"], "type": "object"}, "Site to monitor": {"properties": {"last_checked": {"description": "Last update", "format": "date-time", "type": "string"}, "last_status": {"description": "Last known status", "type": "string"}, "url": {"description": "URL", "type": "string"}}, "required": ["url"], "type": "object"}, "Site to update": {"properties": {"new_url": {"description": "URL", "type": "string"}, "url": {"description": "URL", "type": "string"}}, "required": ["new_url", "url"], "type": "object"}}, "info": {"description": "Site monitoring tool for test assigment", "title": "Sleepless API", "version": "0.1"}, "paths": {"/monitoring/": {"get": {"operationId": "get_monitoring", "parameters": [{"description": "An optional fields mask", "format": "mask", "in": "header", "name": "X-Fields", "type": "string"}], "responses": {"200": {"description": "Success", "schema": {"items": {"$ref": "#/definitions/Site to monitor"}, "type": "array"}}}, "summary": "Returns list of sites with current status", "tags": ["monitoring"]}}, "/sites/": {"delete": {"description": "* Send a JSON object with the URL in request body.\n```\n{\n  \"url\": \"Current URL\"\n}\n```\nReturns list of sites.", "operationId": "delete_sites_collection", "parameters": [{"in": "body", "name": "payload", "required": true, "schema": {"$ref": "#/definitions/Site"}}], "responses": {"204": {"description": "Site successfully deleted from monitoring."}}, "summary": "Deletes site from monitoring", "tags": ["sites"]}, "get": {"operationId": "get_sites_collection", "parameters": [{"description": "An optional fields mask", "format": "mask", "in": "header", "name": "X-Fields", "type": "string"}], "responses": {"200": {"description": "Success", "schema": {"items": {"$ref": "#/definitions/Site"}, "type": "array"}}}, "summary": "Returns list of sites", "tags": ["sites"]}, "post": {"description": "* Send a JSON object with the URL in request body.\n```\n{\n  \"url\": \"New URL\"\n}\n```\nReturns list of sites.", "operationId": "post_sites_collection", "parameters": [{"in": "body", "name": "payload", "required": true, "schema": {"$ref": "#/definitions/Site"}}], "responses": {"201": {"description": "URL successfully added."}}, "summary": "Adds new site to monitor", "tags": ["sites"]}, "put": {"description": "* Send a JSON object with the new and old URL in request body.\n```\n{\n  \"url\": \"Current URL\",\n  \"new_url\": \"New URL\"\n}\n```\nReturns list of sites.", "operationId": "put_sites_collection", "parameters": [{"in": "body", "name": "payload", "required": true, "schema": {"$ref": "#/definitions/Site to update"}}], "responses": {"204": {"description": "Site successfully updated."}}, "summary": "Updates site url", "tags": ["sites"]}}}, "produces": ["application/json"], "responses": {"MaskError": {"description": "When any error occurs on mask"}, "ParseError": {"description": "When a mask can't be parsed"}}, "swagger": "2.0", "tags": [{"description": "Operations with sites", "name": "sites"}, {"description": "Monitoring operations", "name": "monitoring"}]}